// src/app/customer/models/catalogs/customer-analista.ts
export interface Analista {
    id_analista: string;
    id_analista_cve: string;
    id_centro_financiero?: string;
    id_tipo_area_bancaria?: string;
    primer_nombre?: string;
    segundo_nombre?: string;
    apellido_paterno?: string;
    apellido_materno?: string;
    correo?: string;
    proc_pm_banco?: string;
  }
  